src/pyproject.toml
[project]
name = "pipeline-orchestrator"
version = "0.1.0"
description = "Fully bootstrap and orchestrate your infrastructure though a single yaml pipeline"
requires-python = ">=3.13"
dependencies = [
    "pulumi>=3.0.0,<4.0.0",
    "pydantic>=2.10.4",
    "pyyaml>=6.0.2",
]
